Outgoing queues
This is a experimental branch which tries out rq queues. Our current setup doesn't scale if too many receiver are invovled. In this case the dispatcher is not able to cope with it. The main idea is to put sending jobs to a queue (redis backed) and let workers pick from those queues. Right now only one Queue is implemented.